+++ src/pages/party/explore/Explore.tsx
@@ -14,13 +14,11 @@ import { List } from "@/components/list";
 import { Stack } from "@/components/stack";
 import { Tabs } from "@/components/tabs";
 import { Typo } from "@/components/typo";
-import { PartyCard, PartyCardSkeleton } from "@/pages/components";
+import usePartyCardRenderer from "@/hooks/usePartyCardRenderer";
 import {
   useFetchAllParty,
   useFetchJoinedParty,
 } from "@/services/internal/party/query";
-import { PartyInfo } from "@/services/internal/types";
-import { calculateDday } from "@/utils/calculateDday";
 
 const NAME = {
   FIRST: "짠 모임",
@@ -53,6 +51,7 @@ const Explore = () => {
     personnelLoe,
     ageTag,
   });
+
   const { data: allPartyResponse, isLoading: isLoadingAllParty } =
     useFetchAllParty();
   const { data: joinedPartyResponse, isLoading: isLoadingJoinedParty } =
@@ -62,42 +61,11 @@ const Explore = () => {
     ? filteredPartyList
     : allPartyResponse && allPartyResponse.data;
 
-  const handleDday = (date: string) => {
-    const [day] = date.split(" ");
-    return calculateDday(day);
-  };
-
   const isLoadingParty =
     (isFilteredPage && isLoadingFilteredParty) ||
     (!isFilteredPage && isLoadingAllParty);
 
-  const renderPartyCardSkeletons = () =>
-    Array(5)
-      .fill(0)
-      .map((_, index) => <PartyCardSkeleton key={index} />);
-
-  const renderPartyList = (
-    isLoading: boolean,
-    partyData: PartyInfo[] | undefined,
-  ) => {
-    if (isLoading) return renderPartyCardSkeletons();
-
-    if (partyData && partyData.length > 0) {
-      return partyData.map(partyInfo => (
-        <Link to={`/party-detail/${partyInfo.id}`} key={partyInfo.id}>
-          <PartyCard
-            title={partyInfo.title}
-            date={partyInfo.partyDate}
-            partyImage={partyInfo.thumbnail}
-            dDay={handleDday(partyInfo.partyDate)}
-            contributorsAvatars={partyInfo.joinUser.map(user => user.profile)}
-          />
-        </Link>
-      ));
-    } else {
-      return "모임을 찾을 수 없습니다.";
-    }
-  };
+  const renderPartyList = usePartyCardRenderer();
 
   return (
     <Layout
